Added authorization to the delete task

+++ app/Http/Controllers/TaskController.php
@@ -101,6 +101,7 @@ class TaskController extends Controller
      */
     public function destroy(Task $task)
     {
+        Gate::authorize('delete', $task);
         $task->delete();
 
         return redirect(route('tasks.index'));
